# frontend/apps/admin/openapi/7dpanel.v1.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "7DPanel API",
    "description": "Runtime API for 7DPanel 0.1.0.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/server-configuration": {
      "get": {
        "tags": [
          "ServerConfiguration"
        ],
        "operationId": "ServerConfiguration_Get",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerConfigurationSnapshotResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/server-configuration/{key}": {
      "put": {
        "tags": [
          "ServerConfiguration"
        ],
        "operationId": "ServerConfiguration_Put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateServerConfigurationHttpRequest"
                  }
                ]
              }
            }
          },
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerConfigurationUpdateResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServerConfigurationSnapshotResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "string"
          },
          "readAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServerConfigurationFieldResponse"
            }
          }
        }
      },
      "ServerConfigurationFieldResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "group": {
            "type": "string"
          },
          "valueType": {
            "type": "string"
          },
          "editable": {
            "type": "boolean"
          },
          "sensitive": {
            "type": "boolean"
          },
          "isSet": {
            "type": "boolean"
          },
          "restartRequired": {
            "type": "boolean"
          },
          "allowedValues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minimum": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "maximum": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
      "ServerConfigurationUpdateResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "version": {
            "type": "string"
          },
          "savedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "restartRequired": {
            "type": "boolean"
          }
        }
      },
      "UpdateServerConfigurationHttpRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Opaque Bearer access token.",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  }
}